Remove firebase directory and rename error file
