+++ backend/src/errors/index.ts
@@ -1,3 +1,3 @@
 export * from "./errors";
 export * from "./internal";
-export * from "./user";
+export * from "./validation";
